% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{cfnorm}
\alias{cfnorm}
\title{Normal Characteristic Function}
\usage{
cfnorm(t, mean = 0, sd = 1)
}
\arguments{
\item{t}{Vector where the characteristic function will be
evaluated}

\item{mean}{Mean of Laplace distribution}

\item{sd}{Standard deviation of Laplace distribution}
}
\value{
A complex vector with characteristic function values
}
\description{
Obtain the normal characteristic function in a vector
}
\author{
Guillermo Basulto-Elias
}

