% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{compute_scale_par}
\alias{compute_scale_par}
\title{Compute scale parameter for error in kdde}
\usage{
compute_scale_par(error_dist, error_smp, k, error_scale_par = NULL)
}
\arguments{
\item{error_dist}{An integer specifying the error distribution: 1
if non-parametric, 2 if Laplace, 3 if normal.}

\item{error_smp}{Vector of sample of errors or differences of
errors.}

\item{k}{An integer specifying what is being provided as sample of
errors: '1' for pure sample of errors and otherwise for
differences of errors.}

\item{error_scale_par}{NULL by default. It is is given, the
function does nothing but return this value.}
}
\value{
The (positive) standard deviation
}
\description{
This function computes the maximum likelihood estimator of the
standard deviation for the error distribution provided as long as
this error is normal or Laplace and the sample is either a sample
or errors or a sample of differeces of errors (which arises in
panel data).
}
\details{
This function also computes the standard deviation when the error
is approximated with the enpirical charcteristic function, since
it is useful for bandwidth selection.
}
\examples{
## Difference of Laplace errors is provided:
sigma <- 5
smp <- rlaplace(n = 100, sd = sigma) - rlaplace(n = 100, sd = sigma)
kerdec:::compute_scale_par(2, smp, 2)
}
\author{
Guillermo Basulto-Elias
}
