% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecf_im}
\alias{ecf_im}
\title{Imaginary part of empirical characteristic function}
\usage{
ecf_im(t, smp)
}
\arguments{
\item{t}{mxd matrix where the function will be evaluated.}

\item{smp}{nxd matrix with sample size if size n.}
}
\value{
A vector of size m with the imaginary part of the
empirical characteristic function.
}
\description{
Imaginary part of empirical characteristic function of a
d-dimensional random variable. This function is evaluated at m
vectors of size d.
}
\details{
This function must receive matrices. Vectors or values are not
accepted.
}

