% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecf_mod_cpp}
\alias{ecf_mod_cpp}
\title{Modulus of empirical characteristic function}
\usage{
ecf_mod_cpp(t, smp)
}
\arguments{
\item{t}{mxd matrix where the function will be evaluated.}

\item{smp}{nxd matrix with sample size if size n.}
}
\value{
A complex of size m with the modulus of the
empirical characteristic function.
}
\description{
Modulus of empirical characteristic function of a d-dimensional
random variable. This function is evaluated at m vectors of size
d.
}
\details{
This function must receive matrices. Vectors or values are not
accepted.
}

