% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_kernel.R
\name{ft_kernel}
\alias{ft_kernel}
\title{Fourier transform of kernel}
\usage{
ft_kernel(t, ker)
}
\arguments{
\item{t}{Matri}

\item{ker}{}

\item{mat}{nxp matrix or vector of size n.}
}
\value{
A vector of size n.
}
\description{
Compute Fourier transform of product kernels.
}
\details{
The Fourier transform of all the kernel functions considered for
this package have a bounded support, as it is usual in kernel
deconvolution methods.
}

