% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{laplace_convol_loglik}
\alias{laplace_convol_loglik}
\title{Log-likelihood of Laplace Convolution}
\usage{
laplace_convol_loglik(sigma, smp)
}
\arguments{
\item{sigma}{Value or vector of positive values where the this
function will be evaluated.}

\item{smp}{Sample of differences of Laplace random variables.}
}
\value{
A vector of values of the log-likelihood.
}
\description{
Negative value of the log-likelihood of the difference of two
independent Laplace random variables.
}
\author{
Guillermo Basulto-Elias
}
