% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{mle_laplace_diffs}
\alias{mle_laplace_diffs}
\title{MLE of the std. dev. for Laplace convolution}
\usage{
mle_laplace_diffs(smp)
}
\arguments{
\item{smp}{Sample of differences of Laplace random variables.}
}
\value{
The MLE of the standard deviation
}
\description{
Compute the maximum likelihood estimator for the standard
deviation of the difference of two independent Laplace random
variables.
}
\author{
Guillermo Basulto-Elias
}

