% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerdec_dens.R
\name{optimize_bw}
\alias{optimize_bw}
\title{Find positive minimizer with the input function}
\usage{
optimize_bw(f, h0, ...)
}
\arguments{
\item{f}{Function to be minimized. Tipically a version of CV, MISE
or AMISE.}

\item{h0}{Initial bandwidth value.}

\item{...}{Arguments passed to f besides the first argument, which
must be the bandwidth.}
}
\value{
The output of the nlm function.
}
\description{
Most methods used to find optimal bandwidths require to minimize a
function (tipically MISE, AMISE, CV, ...). The argument ought to
be positive. The fastest method that I empirically found was
nlm. If the function is evaluated at a negative number or the
output is not finite or not a number, the function will return the
maximum possible number, so nlm can work better with the input
function.
}
\examples{
\dontrun{
## Bad initial point (it is negative).
kerdec:::optimize_bw(function(x) x^2, -5)
}

## Good initial point.
kerdec:::optimize_bw(function(x) x^2, 0.5)
}
\author{
Guillermo Basulto-Elias
}

