% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{plaplace}
\alias{plaplace}
\title{Laplace Cumulative Distribution Function}
\usage{
plaplace(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{Vector where the cdf will be evaluated}

\item{mean}{Mean of Laplace distribution}

\item{sd}{Standard deviation of Laplace distribution}
}
\value{
A vector with cdf values
}
\description{
Evaluate a Laplace cdf in a vector
}
\examples{
x <- seq(-2, 4, 0.05)
vals <- plaplace(x, mean = 1, sd = 0.99)
plot(x, vals, type = "l")
}
\author{
Guillermo Basulto-Elias
}
\seealso{
\code{\link{dlaplace}} for probability density function
    and \code{\link{rlaplace}} for random number generation.
}

