% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_differences}
\alias{process_differences}
\title{Process differences for panel data}
\usage{
process_differences(smp, method)
}
\description{
Panel data allows to approximate the characteristic function of
the error taking differences of observations for each
individual. This way is not unique when there are more than two
replicates per individual. This function allows to do it in
several ways.
}

