% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{rlaplace}
\alias{rlaplace}
\title{Laplace Random Generator}
\usage{
rlaplace(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{Sample size}

\item{mean}{Mean of Laplace distribution}

\item{sd}{Standard deviation of Laplace distribution}
}
\value{
A vector with random values
}
\description{
Generate Laplace random sample
}
\author{
Guillermo Basulto-Elias
}

