% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_distributions.R
\name{dlaplace}
\alias{dlaplace}
\title{Laplace Probability Density Function}
\usage{
dlaplace(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{Vector where the pdf will be evaluated}

\item{mean}{Mean of Laplace distribution}

\item{sd}{Standard deviation of Laplace distribution}
}
\value{
A vector with pdf values
}
\description{
Evaluate a Laplace pdf in a vector
}
\examples{
x <- seq(-2, 4, 0.05)
vals <- dlaplace(x, mean = 1, sd = 0.99)
plot(x, vals, type = "l")
}
\author{
Guillermo Basulto-Elias
}
\seealso{
\code{\link{plaplace}} for cumulative distribution
    function, \code{\link{rlaplace}} for random number generation.
}

