% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_kernel.R
\name{ft_kernel}
\alias{ft_kernel}
\title{Fourier transform of kernel}
\usage{
ft_kernel(t, ker)
}
\arguments{
\item{t}{A vector of size n of a nxd matrix where the Fourier
transform of a kernel or a d-product kernel (if t is a matrix)
will be evaluated.}

\item{ker}{Character or number specifying the kernel. All the
kernels are symmetric and they have a FT supported on [-1, 1].
The possible choices
are:
\itemize{
  \item{1 or "sinc"}{ Sinc kernel}
  \item{2 or "triangular"}{ Kernel whose FT is a triangle}
  \item{3 or "triweight"}{ Kernel whose FT is proportional to
                          triweight kernel}
  \item{4 or "tricube"}{ Kernel whose FT is proportional to
                          tricube kernel}
  \item{5 or "flat-top"}{ A kernel with FT equal to one around
                         zero and decreasing linearly to zero
                         at -1/2 and 1/2 }}}
}
\value{
A vector of size n.
}
\description{
Evaluate Fourier transform of kernels and product kernels.
}
\details{
The Fourier transform (FT) of all the kernel functions considered
for this package have a bounded support, as it is usual in kernel
deconvolution methods. All the kernels are symmetric.

See vignette for detailed examples.
}
\examples{
t <- seq(-1.5, 1.5, 0.05)
vals <- ft_kernel(t, "flat-top")
plot(t, vals, type = "l")
}
