% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerdec_dens (Guillermo Basulto Elías's conflicted copy 2016-10-17).R, R/kerdec_dens.R
\name{get_sampling_scenario}
\alias{get_sampling_scenario}
\title{Identify sampling scenario in kerdec}
\usage{
get_sampling_scenario(k, error_dist, error_scale_par)

get_sampling_scenario(k, error_dist, error_scale_par)
}
\description{
Based on the given parameters to deconvolution function,
determines the sampling scenario and displays a message.

Based on the given parameters to deconvolution function,
determines the sampling scenario and displays a message.
}

