% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_differences.R
\name{process_differences}
\alias{process_differences}
\title{Process differences for panel data}
\usage{
process_differences(smp, method)
}
\arguments{
\item{smp}{n x d matrix}

\item{method}{Integer or word specifying method to process
differences.
   1 or "all", all pairwise differences.
   2 or "vs_first", all minus first.
   3 or "indep_pairs", independent columns.}
}
\description{
Panel data allows to approximate the characteristic function of
the error taking differences of observations for each
individual. This way is not unique when there are more than two
replicates per individual. This function allows to do it in
several ways.
}

