% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bw.R
\name{select_bw}
\alias{select_bw}
\title{Bandwidth Selection}
\usage{
select_bw(smp, method = c("CV", "NR")[1], kernel = "flat", h0 = NULL,
  error_smp = NULL, error_dist = "None", error_scale_par = NULL,
  resolution = 128, error_proc = "all", panel_proc = c("keep_first",
  "take_aver")[1], truncation_bound = NULL)
}
\arguments{
\item{smp}{It is either a vector of size n or a nxl matrix (for
repeated observations; l per each individual).}

\item{method}{Method used to select the bandwidth. The currently
available methods are "CV" (cross-validation) and "NR" (normal
reference). Cross-validation is the default.}

\item{kernel}{Kernel whose Fourier transform has bounded
support. There are currently 5 kernel programmed: "sinc",
"vp", "triw", "tric" and "flat". See vignette for further
details and examples. Flat-top kernel is the default.}

\item{h0}{Optional argument used as initial value to look for the
optimal value.}

\item{error_smp}{Optional vector errors. It is necessary to
approximate the error distribution if it is unknown.}

\item{error_dist}{Three possible values are accepted. c("Normal",
"Laplace", "None") fcwef}
}
\value{
A list
}
\description{
This function provides a bandwidth for kernel denvolvolution
density estimator. This works for several deconvolution scenarios,
specifically, when the error distribution is known, when a sample
of pure errors is available instead and when a contaminated sample
is available as panel data.
}
\details{
See the vignette for more details.
}
\author{
Guillermo Basulto-Elias
}

